+++ Sources/StreamChatSwiftUI/ChatMessageList/Gallery/MediaViewer.swift
@@ -194,6 +194,10 @@ struct StreamVideoPlayer: View {
         utils.fileCDN
     }
 
+    private var avPlayerProvider: AVPlayerProvider {
+        utils.avPlayerProvider
+    }
+
     let url: URL
 
     @State var avPlayer: AVPlayer?
@@ -221,6 +225,16 @@ struct StreamVideoPlayer: View {
                     avPlayer = AVPlayer(url: url)
                     try? AVAudioSession.sharedInstance().setCategory(.playback, options: [])
                     avPlayer?.play()
+                    self.avPlayerProvider.player(for: url) { result in
+                        switch result {
+                        case let .success(player):
+                            self.avPlayer = player
+                            try? AVAudioSession.sharedInstance().setCategory(.playback, options: [])
+                            self.avPlayer?.play()
+                        case let .failure(error):
+                            self.error = error
+                        }
+                    }
                 case let .failure(error):
                     self.error = error
                 }