+++ StreamChatSwiftUITests/Infrastructure/Mocks/VideoPreviewLoader_Mock.swift
@@ -2,6 +2,7 @@
 // Copyright © 2026 Stream.io Inc. All rights reserved.
 //
 
+@testable import StreamChat
 @testable import StreamChatSwiftUI
 import UIKit
 import XCTest
@@ -9,6 +10,7 @@ import XCTest
 /// Mock implementation of `VideoPreviewLoader`.
 class VideoPreviewLoader_Mock: VideoPreviewLoader {
     var loadPreviewVideoCalled = false
+    var loadPreviewVideoWithAttachmentCalled = false
 
     func loadPreviewForVideo(at url: URL, completion: @escaping @MainActor (Result<UIImage, Error>) -> Void) {
         loadPreviewVideoCalled = true
@@ -16,5 +18,14 @@ class VideoPreviewLoader_Mock: VideoPreviewLoader {
         StreamConcurrency.onMain {
             completion(.success(ImageLoader_Mock.defaultLoadedImage))
         }
+    }
+
+    func loadPreviewForVideo(
+        with attachment: ChatMessageVideoAttachment,
+        completion: @escaping (Result<UIImage, Error>) -> Void
+    ) {
+        loadPreviewVideoWithAttachmentCalled = true
+
+        completion(.success(ImageLoader_Mock.defaultLoadedImage))
     }
 }